IGNITE-23359 Sql. Additional tests for "out of range" arithmetic operations
