+++ modules/sql-engine/src/integrationTest/java/org/apache/ignite/internal/sql/engine/ItDynamicParameterTest.java
@@ -20,6 +20,7 @@ package org.apache.ignite.internal.sql.engine;
 import static org.apache.ignite.internal.lang.IgniteStringFormatter.format;
 import static org.apache.ignite.internal.sql.engine.util.SqlTestUtils.assertThrowsSqlException;
 import static org.apache.ignite.internal.testframework.IgniteTestUtils.await;
+import static org.apache.ignite.lang.ErrorGroups.Sql.RUNTIME_ERR;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
@@ -480,6 +481,41 @@ public class ItDynamicParameterTest extends BaseSqlIntegrationTest {
             QueryProcessor qryProc = queryProcessor();
             await(qryProc.prepareSingleAsync(properties, null, stmt.toString())).parameterTypes();
         });
+    }
+
+    @ParameterizedTest(name = "{1} {2}")
+    @MethodSource("decimalOverflows")
+    @SuppressWarnings("ThrowableNotThrown")
+    public void testCalcOpOverflow(SqlTypeName type, String expr, Object param) {
+        assertThrowsSqlException(RUNTIME_ERR, type.getName() + " out of range", () -> sql(expr, param));
+    }
+
+    private static Stream<Arguments> decimalOverflows() {
+        return Stream.of(
+                // BIGINT
+                arguments(SqlTypeName.BIGINT, "SELECT -(?)", -9223372036854775808L),
+                arguments(SqlTypeName.BIGINT, "SELECT CAST(-? AS BIGINT)/-1", "9223372036854775808"),
+                arguments(SqlTypeName.BIGINT, "SELECT CAST(-? AS BIGINT) * -1", "9223372036854775808"),
+                arguments(SqlTypeName.BIGINT, "SELECT (?::BIGINT/-1)::BIGINT", "-9223372036854775808"),
+
+                // INTEGER
+                arguments(SqlTypeName.INTEGER, "SELECT -(?)", -2147483648),
+                arguments(SqlTypeName.INTEGER, "SELECT CAST(-? AS INTEGER)/-1", "2147483648"),
+                arguments(SqlTypeName.INTEGER, "SELECT CAST(-? AS INTEGER) * -1", "2147483648"),
+                arguments(SqlTypeName.INTEGER, "SELECT (?/-1)::INTEGER", "-2147483648"),
+
+                // SMALLINT
+                arguments(SqlTypeName.SMALLINT, "SELECT -CAST(? AS SMALLINT)", -32768),
+                arguments(SqlTypeName.SMALLINT, "SELECT (CAST(-? AS SMALLINT)/-1)::SMALLINT", 32768),
+                arguments(SqlTypeName.SMALLINT, "SELECT (CAST(-? AS SMALLINT) * -1)::SMALLINT", 32768),
+                arguments(SqlTypeName.SMALLINT, "SELECT (?/-1)::SMALLINT", -32768),
+
+                // TINYINT
+                arguments(SqlTypeName.TINYINT, "SELECT -CAST(? AS TINYINT)", -128),
+                arguments(SqlTypeName.TINYINT, "SELECT (CAST(-? AS TINYINT)/-1)::TINYINT", 128),
+                arguments(SqlTypeName.TINYINT, "SELECT (CAST(-? AS TINYINT) * -1)::TINYINT", 128),
+                arguments(SqlTypeName.TINYINT, "SELECT (?/-1)::TINYINT", -128)
+        );
     }
 
     @Override
