+++ modules/sql-engine/src/integrationTest/sql/insert/test_insert_type.test
@@ -30,3 +30,239 @@ SELECT * FROM strings ORDER BY cast(a AS INTEGER)
 4
 13
 NULL
+
+statement ok
+CREATE TABLE t8(id int, i TINYINT);
+
+statement ok
+CREATE TABLE t16(id int, i SMALLINT);
+
+statement ok
+CREATE TABLE t32(id int, i INTEGER);
+
+statement ok
+CREATE TABLE t64(id int, i BIGINT);
+
+statement ok
+INSERT INTO t8 VALUES(1, -128);
+
+statement ok
+INSERT INTO t16 VALUES(1, -32768);
+
+statement ok
+INSERT INTO t32 VALUES(1, -2147483648);
+
+statement ok
+INSERT INTO t64 VALUES(1, -9223372036854775808);
+
+#Multiply
+
+statement error
+INSERT INTO t16 VALUES(2, (SELECT i * i FROM t8 WHERE id=1));
+
+statement error
+INSERT INTO t32 VALUES(2, (SELECT i * i FROM t16 WHERE id=1));
+
+statement error
+INSERT INTO t64 VALUES(2, (SELECT i * i FROM t32 WHERE id=1));
+
+
+for val in [SMALLINT, INT, DECIMAL(10\,0), BIGINT, REAL, FLOAT, DOUBLE]
+
+statement ok
+INSERT INTO t16 VALUES(2, (SELECT i * i::${val} FROM t8 WHERE id=1));
+
+query T
+SELECT i FROM t16 WHERE id=2;
+----
+16384
+
+statement ok
+DELETE FROM t16 WHERE id=2;
+
+endfor
+
+
+for val in [INT, DECIMAL(10\,0), BIGINT, REAL, FLOAT, DOUBLE]
+
+statement ok
+INSERT INTO t32 VALUES(2, (SELECT i * i::${val} FROM t16 WHERE id=1));
+
+query T
+SELECT i FROM t32 WHERE id=2;
+----
+1073741824
+
+statement ok
+DELETE FROM t32 WHERE id=2;
+
+endfor
+
+
+for val in [DECIMAL(19\,0), BIGINT, REAL, FLOAT, DOUBLE]
+
+statement ok
+INSERT INTO t64 VALUES(2, (SELECT i * i::${val} FROM t32 WHERE id=1));
+
+query T
+SELECT i FROM t64 WHERE id=2;
+----
+4611686018427387904
+
+statement ok
+DELETE FROM t64 WHERE id=2;
+
+endfor
+
+
+#Divide
+
+for val in [TINYINT, SMALLINT, INT, DECIMAL(10\,0), BIGINT, REAL, FLOAT, DOUBLE]
+
+statement ok
+INSERT INTO t16 VALUES(2, (SELECT i::${val}/-1 FROM t8 WHERE id=1));
+
+query T
+SELECT i FROM t16 WHERE id=2;
+----
+128
+
+statement ok
+DELETE FROM t16 WHERE id=2;
+
+endfor
+
+
+for val in [SMALLINT, INT, DECIMAL(10\,0), BIGINT, REAL, FLOAT, DOUBLE]
+
+statement ok
+INSERT INTO t32 VALUES(2, (SELECT i::${val}/-1 FROM t16 WHERE id=1));
+
+query T
+SELECT i FROM t32 WHERE id=2;
+----
+32768
+
+statement ok
+DELETE FROM t32 WHERE id=2;
+
+endfor
+
+
+for val in [DECIMAL(19\,0), BIGINT, REAL, FLOAT, DOUBLE]
+
+statement ok
+INSERT INTO t64 VALUES(2, (SELECT i::${val}/-1 FROM t32 WHERE id=1));
+
+query T
+SELECT i FROM t64 WHERE id=2;
+----
+2147483648
+
+statement ok
+DELETE FROM t64 WHERE id=2;
+
+endfor
+
+
+
+#Add
+
+for val in [SMALLINT, INT, DECIMAL(10\,0), BIGINT, REAL, FLOAT, DOUBLE]
+
+statement ok
+INSERT INTO t16 VALUES(2, (SELECT i::${val} + i FROM t8 WHERE id=1));
+
+query T
+SELECT i FROM t16 WHERE id=2;
+----
+-256
+
+statement ok
+DELETE FROM t16 WHERE id=2;
+
+endfor
+
+
+for val in [INT, DECIMAL(10\,0), BIGINT, REAL, FLOAT, DOUBLE]
+
+statement ok
+INSERT INTO t32 VALUES(2, (SELECT i::${val} + i FROM t16 WHERE id=1));
+
+query T
+SELECT i FROM t32 WHERE id=2;
+----
+-65536
+
+statement ok
+DELETE FROM t32 WHERE id=2;
+
+endfor
+
+
+for val in [DECIMAL(11\,0), BIGINT, REAL, FLOAT, DOUBLE]
+
+statement ok
+INSERT INTO t64 VALUES(2, (SELECT i::${val} + i FROM t32 WHERE id=1));
+
+query T
+SELECT i FROM t64 WHERE id=2;
+----
+-4294967296
+
+statement ok
+DELETE FROM t64 WHERE id=2;
+
+endfor
+
+
+#Subtract
+
+for val in [SMALLINT, INT, DECIMAL(10\,0), BIGINT, REAL, FLOAT, DOUBLE]
+
+statement ok
+INSERT INTO t16 VALUES(2, (SELECT i::${val} - 1 FROM t8 WHERE id=1));
+
+query T
+SELECT i FROM t16 WHERE id=2;
+----
+-129
+
+statement ok
+DELETE FROM t16 WHERE id=2;
+
+endfor
+
+
+for val in [INT, DECIMAL(10\,0), BIGINT, REAL, FLOAT, DOUBLE]
+
+statement ok
+INSERT INTO t32 VALUES(2, (SELECT i::${val} - 1 FROM t16 WHERE id=1));
+
+query T
+SELECT i FROM t32 WHERE id=2;
+----
+-32769
+
+statement ok
+DELETE FROM t32 WHERE id=2;
+
+endfor
+
+
+for val in [DECIMAL(11\,0), BIGINT, DOUBLE]
+
+statement ok
+INSERT INTO t64 VALUES(2, (SELECT i::${val} - 1 FROM t32 WHERE id=1));
+
+query T
+SELECT i FROM t64 WHERE id=2;
+----
+-2147483649
+
+statement ok
+DELETE FROM t64 WHERE id=2;
+
+endfor
+
+
